feat(db schema): simplify database schema
Express tags, authors, platform, and type as a property in the item table instead of each having their own table and creating a joint table with items for each.

This makes it easier to enable the AI agent to interact with the database because of the simplicity.
Also, the complexity and performance hit from many joint tables wasn't justified.